% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pacf2phi.R
\name{ar.pacf2phi}
\alias{ar.pacf2phi}
\title{Convert partial autocorrelation coefficients to autoregressive coeffients}
\usage{
ar.pacf2phi(pac)
}
\arguments{
\item{pac}{partial autocorrelation coefficient (vector)}
}
\description{
Convert partial autocorrelation coefficients to autoregressive coeffients
}
\keyword{internal}
