% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pacf2phi.R
\name{ar.phi2pacf}
\alias{ar.phi2pacf}
\title{Convert autoregressive coeffients to partial autocorrelation coefficients}
\usage{
ar.phi2pacf(phi)
}
\arguments{
\item{pac}{autoregressive coefficients (vector)}
}
\description{
Convert autoregressive coeffients to partial autocorrelation coefficients
}
\keyword{internal}
