% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biascorrect.R
\name{biascorrect}
\alias{biascorrect}
\title{AR Bias Correction for short time series}
\usage{
biascorrect(
  phi = NULL,
  n = NULL,
  method = c("yw", "mle", "burg", "cmle"),
  order = NULL,
  x = NULL
)
}
\arguments{
\item{phi}{single value (AR1) or two-dimensional vector (AR2) containing the AR estimates subject to bias correction. Not required
if time series \code{x} is used as input.}

\item{n}{integer for the length of the time series. Needs to be between 10 and 50. Not required
if time series \code{x} is used as input.}

\item{method}{Character string specifying the method used to estimate the autoregressive
coefficients. Needs to be either \code{mle}, \code{cmle}, \code{burg} or \code{yw},
specifying the exact MLE, the conditional MLE, Burg's method and the Yule-Walker solution,
respectively.}

\item{order}{order of the estimated autoregressive process. \code{x} is used as input.}

\item{x}{The time series to be fitted by AR(1) or AR(2) in which \code{order} needs to be specified.
The default estimation method is the exact MLE if none is specified.}
}
\value{
A list with the following elements
\itemize{
\item \code{phi.hat} The original estimates of the AR coefficients
\item \code{phi.correct} The bias-corrected estimates
\item \code{ci.hat} The 95\% confidence interval for the original estimates
\item \code{ci.correct} The 95\% confidence interval for the bias-corrected estimates
}
}
\description{
Gives bias-corrected estimates and 95\% confidence intervals for autoregressive coefficients of  AR(1) and AR(2) processes,
for sample sizes n=10,11, \ldots, 50.
}
\examples{
# pre-computed estimates
biascorrect(phi=c(0.5,-0.2), n=15, method="mle")


\donttest{
# simulating series with ARIMA sim
series <- arima.sim(n = 20, list(ar = c(0.2, 0.5), sd = sqrt(0.04)))
biascorrect(x=series, order=2, method="mle")
}
}
\references{
Sørbye, S. H., Nicolau, P. G. & Rue, H. (2021). Finite-sample properties of estimators for first
and second order autoregressive processes.
}
