% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_g.R
\name{eval.g.ar2}
\alias{eval.g.ar2}
\title{Evaluate sum of Hermite polynomials for g(pacf1) and g(pacf2)}
\usage{
eval.g.ar2(x, beta)
}
\description{
Evaluate sum of Hermite polynomials for g(pacf1) and g(pacf2)
}
\keyword{internal}
