% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_g.R
\name{eval.g.sn}
\alias{eval.g.sn}
\title{Evaluate separate regression model for each of the SN-parameters + rho}
\usage{
eval.g.sn(x, beta)
}
\description{
Evaluate separate regression model for each of the SN-parameters + rho
}
\keyword{internal}
