% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcopula.R
\name{f.copula}
\alias{f.copula}
\title{Find samples using Copula approach to preserve correlation}
\usage{
f.copula(theta.vec)
}
\description{
Find samples using Copula approach to preserve correlation
}
\keyword{internal}
