% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.orig.R
\name{f.orig}
\alias{f.orig}
\title{Find original estimate for coefficients if time series is given}
\usage{
f.orig(x, method, p)
}
\description{
Find original estimate for coefficients if time series is given
}
\keyword{internal}
