% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_order.R
\name{f_order}
\alias{f_order}
\title{AR(1) and AR(2): Defining map etc for K=3 order regression}
\usage{
f_order(p)
}
\description{
AR(1) and AR(2): Defining map etc for K=3 order regression
}
\keyword{internal}
